requirements.txt
pyautogui
setuptools
pyyaml
webdriver-manager==4.0.2